reciclamJA/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
coreapi==2.3.3
coreschema==0.0.4
Django==5.1.6
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
idna==3.10
itypes==1.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
mysqlclient==2.2.7
PyJWT==2.9.0
requests==2.32.3
setuptools==75.8.2
sqlparse==0.5.3
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
